x/weightslice: add weightslice package
This commit adds a new package `weightslice` that provides a "weighted"
slice. It is a wrapper around a slice of elements where each element
has an assigned weight that impacts the order of iteration, supporting
both ascending and descending order.

I wanted to use this for selecting DNS upstreams based on round-trip
time.

Signed-off-by: David Bond <davidsbond93@gmail.com>