x/weightslice: use entry identifiers
This commit modifies things under the hood for the `weightslice` package.

Each element is now assigned an identifier that matches their starting
index of when they were appended. This gives a stable identifier with
which we can always reference.

This is important as in concurrent scenarios the range is being performed
over a snapshot of the underlying slice and indexes may change due to
calls to sort in other goroutines. Using a fixed identifier means that
calls to `SetWeight` will always refer to a specific element, rather
than whatever element lives at that index.

Signed-off-by: David Bond <davidsbond93@gmail.com>